controller: Move manager and indexers to dedicated files
